third_party/WebKit/LayoutTests/http/tests/local/formdata/send-form-data-with-bad-string-expected.txt

CONSOLE ERROR: line 20: Uncaught SyntaxError: Illegal return statement
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
